+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.3.1",
     "react-inlinesvg": "^4.1.3",
     "react-redux": "^9.1.2",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
     "recharts": "^2.13.0-alpha.5",
     "styled-components": "^6.1.13",
