+++ src/routes/ordenes/ordenes.component.tsx
@@ -2,10 +2,10 @@ import moment from "moment";
 import { Fragment, useState } from "react";
 import 'moment/locale/es';
 import {ReactComponent as SortIcon} from '@material-design-icons/svg/outlined/sort.svg';
-import CarouselWithButtons from "../carousel-with-buttons/carousel-with-buttons.component";
+import CarouselWithButtons from "../../components/carousel-with-buttons/carousel-with-buttons.component";
 import { mesesList, ordenesAnosList, ordenesList } from "../../utils/constantes-test.utils";
 
-const OrderList = () => {
+const Ordenes = () => {
     
     const [ordenesAnos] = useState(ordenesAnosList);
     const [ordenes] = useState(ordenesList);
@@ -99,4 +99,4 @@ const OrderList = () => {
     )
 }
 
-export default OrderList;
+export default Ordenes;
